+++ src/backend/resolvers/queries/organizations/mbm/lbm/index.ts
@@ -2,7 +2,7 @@ import lbmOrganizations from './lbm'
 import lbmServices from './service'
 import lbmServicesCategories from './serviceCategory'
 import lbmTypes from './type'
-// import lbmKeyEvents from './keyEvents'
+import lbmKeyEvents from './keyEvents'
 import joins from './joins'
 import views from './views'
 
@@ -11,6 +11,7 @@ export default {
   lbmTypes,
   lbmServicesCategories,
   lbmServices,
+  lbmKeyEvents,
   ...joins,
   ...views,
 }