+++ src/backend/resolvers/queries/organizations/mbm/lbm/keyEvents.ts
@@ -1,0 +1,4 @@
+const lbmKeyEvents = async (parent, args, { pulseCoreDb }) =>
+  pulseCoreDb.collection('lbms.keyEvents').find().toArray()
+
+export default lbmKeyEvents