+++ src/backend/typeDefs/mutations/organization/mbm/lbm/relational-typeDefs/index.ts
@@ -2,7 +2,7 @@ import serviceAndServiceCategoryTypeDefs from './serviceAndServiceCategory'
 import lbmAndServiceTypeDefs from './lbmAndService'
 import lbmAndLbmTypeTypeDefs from './lbmAndLbmType'
 // import lbmAndPersonTypeDefs from './lbmAndPerson'
-// import lbmAndKeyEvent from './lbmAndKeyEvent'
+import lbmAndKeyEvent from './lbmAndKeyEvent'
 import lbmAndPayer from './lbmAndPayer'
 
 export default [
@@ -10,6 +10,6 @@ export default [
   ...lbmAndServiceTypeDefs,
   ...lbmAndLbmTypeTypeDefs,
   // ...lbmAndPersonTypeDefs,
-  // ...lbmAndKeyEvent,
+  ...lbmAndKeyEvent,
   ...lbmAndPayer,
 ]